moved interpolation class to a seperate file
